feat(devices): consolidate terminals+fiscal+hardware into branch hub, settings routes redirect
