+++ frontend/src/features/fiscal/FiscalRecoveryPage.tsx
@@ -1,19 +1,8 @@
-import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useAuthStore } from '../../store/authStore';
-import { UserRole } from '../../types';
 import QueryStateGate from '../../components/ui/QueryStateGate';
 import { useFormatCurrencyExtended } from '../../hooks/useFormatCurrency';
 import { useFormatDate } from '../../hooks/useFormatDate';
-import {
-  useListPendingReceipts,
-  useRetryReceipt,
-  useListFiscalDevices,
-  useRegisterFiscalDevice,
-  useRetireFiscalDevice,
-  type FiscalReceipt,
-  type FiscalDevice,
-} from './fiscalApi';
+import { useListPendingReceipts, useRetryReceipt, type FiscalReceipt } from './fiscalApi';
 
 /**
  * Manual recovery panel for fiscal receipts.
@@ -41,8 +30,6 @@ export default function FiscalRecoveryPage({ embedded = false }: { embedded?: bo
           <p className="text-sm text-gray-600">{t('hummytummy.fiscalRecovery.subtitle')}</p>
         </header>
       )}
-
-      <FiscalDevicesPanel />
 
       {/* Refresh lives here (not in the header above) so it stays reachable
           when embedded hides the header. */}
@@ -109,128 +96,6 @@ export default function FiscalRecoveryPage({ embedded = false }: { embedded?: bo
         </table>
       </QueryStateGate>
     </div>
-  );
-}
-
-/**
- * Register / list / retire physical fiscal devices (yazarkasa / ÖKC).
- *
- * HONESTY: this is the create-site that makes the payment-finalizer's
- * yazarkasa-receipt path reachable — but it does NOT make receipts print.
- * A real GMP-3 ÖKC only issues once its certified hardware is wired through
- * the local bridge and the bridge acks. The banner says so plainly so an
- * operator never reads "registered" as "fiscalizing".
- */
-function FiscalDevicesPanel() {
-  const { t } = useTranslation('common');
-  const devicesQuery = useListFiscalDevices();
-  const { data: devices = [] } = devicesQuery;
-  const register = useRegisterFiscalDevice();
-  const retire = useRetireFiscalDevice();
-  // Register / retire are ADMIN-only on the backend; don't show controls that
-  // would just 403 for a MANAGER. MANAGERs still see the device list.
-  const isAdmin = useAuthStore((s) => s.user?.role) === UserRole.ADMIN;
-
-  const [providerId, setProviderId] = useState('fiscal_hugin');
-  const [serial, setSerial] = useState('');
-  const [model, setModel] = useState('');
-
-  const submit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!serial.trim()) return;
-    register.mutate(
-      { providerId, serial: serial.trim(), model: model.trim() || undefined },
-      {
-        onSuccess: () => {
-          setSerial('');
-          setModel('');
-        },
-      },
-    );
-  };
-
-  return (
-    <section className="space-y-3 rounded border p-4">
-      <div>
-        <h2 className="text-lg font-semibold">{t('hummytummy.fiscalDevices.title')}</h2>
-        <p className="mt-1 rounded bg-amber-50 px-3 py-2 text-xs text-amber-800">
-          {t('hummytummy.fiscalDevices.gateNote')}
-        </p>
-      </div>
-
-      {isAdmin && (
-      <form onSubmit={submit} className="flex flex-wrap items-end gap-2">
-        <label className="flex flex-col text-xs">
-          <span className="mb-1 text-gray-600">{t('hummytummy.fiscalDevices.provider')}</span>
-          <select
-            value={providerId}
-            onChange={(e) => setProviderId(e.target.value)}
-            className="rounded border px-2 py-1.5 text-sm"
-          >
-            <option value="fiscal_hugin">Hugin</option>
-            <option value="fiscal_beko">Beko</option>
-            <option value="fiscal_paygo">Paygo (SP630)</option>
-          </select>
-        </label>
-        <label className="flex flex-col text-xs">
-          <span className="mb-1 text-gray-600">{t('hummytummy.fiscalDevices.serial')}</span>
-          <input
-            value={serial}
-            onChange={(e) => setSerial(e.target.value)}
-            className="rounded border px-2 py-1.5 text-sm"
-            placeholder="ÖKC-…"
-          />
-        </label>
-        <label className="flex flex-col text-xs">
-          <span className="mb-1 text-gray-600">{t('hummytummy.fiscalDevices.model')}</span>
-          <input
-            value={model}
-            onChange={(e) => setModel(e.target.value)}
-            className="rounded border px-2 py-1.5 text-sm"
-          />
-        </label>
-        <button
-          type="submit"
-          disabled={register.isPending || !serial.trim()}
-          className="rounded bg-blue-600 px-3 py-1.5 text-sm text-white hover:bg-blue-700 disabled:opacity-50"
-        >
-          {t('hummytummy.fiscalDevices.register')}
-        </button>
-      </form>
-      )}
-
-      <QueryStateGate
-        query={devicesQuery}
-        loading={<div className="text-sm text-gray-500">{t('hummytummy.common.loading')}</div>}
-        isEmpty={devices.length === 0}
-        empty={<div className="text-sm text-gray-500">{t('hummytummy.fiscalDevices.empty')}</div>}
-      >
-        <ul className="divide-y rounded border text-sm">
-          {devices.map((d: FiscalDevice) => (
-            <li key={d.id} className="flex items-center justify-between px-3 py-2">
-              <div>
-                <span className="font-mono text-xs">{d.providerId}</span>
-                <span className="mx-2 text-gray-400">·</span>
-                <span>{d.serial}</span>
-                {d.model && <span className="ml-2 text-gray-500">{d.model}</span>}
-                <span className="ml-2 inline-flex rounded bg-gray-100 px-2 py-0.5 text-xs">
-                  {t(`hummytummy.fiscalDevices.status.${d.status}`, { defaultValue: d.status })}
-                </span>
-              </div>
-              {isAdmin && d.status !== 'retired' && (
-                <button
-                  className="rounded border px-2 py-1 text-xs text-red-700 hover:bg-red-50 disabled:opacity-50"
-                  disabled={retire.isPending}
-                  onClick={() => retire.mutate(d.id)}
-                >
-                  {t('hummytummy.fiscalDevices.retire')}
-                </button>
-              )}
-            </li>
-          ))}
-        </ul>
-      </QueryStateGate>
-    </section>
   );
 }
