+++ frontend/src/i18n/locales/ar/common.json
@@ -489,8 +489,12 @@
       "notFound": "الفرع غير موجود أو ليس لديك صلاحية الوصول.",
       "tabs": {
         "devices": "الأجهزة",
+        "terminals": "أجهزة الدفع",
+        "fiscal": "الصندوق",
+        "hardware": "الطابعة والدرج",
         "network": "الشبكة المحلية"
-      }
+      },
+      "scopeHint": "انتقل إلى هذا الفرع من الشريط العلوي لإدارة الطرفيات وصندوق النقد."
     },
     "network": {
       "tokenTitle": "رمز الجسر (يُعرض مرة واحدة)",
