+++ frontend/src/i18n/locales/ru/common.json
@@ -489,8 +489,12 @@
       "notFound": "Филиал не найден или у вас нет доступа.",
       "tabs": {
         "devices": "Устройства",
+        "terminals": "Платёжные терминалы",
+        "fiscal": "Касса",
+        "hardware": "Принтер и денежный ящик",
         "network": "Локальная сеть"
-      }
+      },
+      "scopeHint": "Чтобы управлять терминалами и кассой, переключитесь на этот филиал в верхней панели."
     },
     "network": {
       "tokenTitle": "Токен моста (показывается один раз)",
