+++ frontend/src/pages/settings/PaymentTerminalsSettingsPage.tsx
@@ -40,7 +40,7 @@ const PROVIDER_LABELS: Record<string, string> = {
 };
 const providerLabel = (id: string) => PROVIDER_LABELS[id] ?? id;
 
-const PaymentTerminalsSettingsPage = () => {
+export const PaymentTerminalsPanel = () => {
   const { t } = useTranslation('settings');
   const formatPrice = useFormatCurrency();
   const { data: providers } = useTerminalProviders();
@@ -121,134 +121,128 @@ const PaymentTerminalsSettingsPage = () => {
   };
 
   return (
-    <div className="h-full overflow-auto p-4 md:p-6">
-      {/* Page header */}
-      <div className="mb-6 flex flex-wrap items-start justify-between gap-3">
-        <div>
-          <h1 className="text-xl font-heading font-bold text-slate-900">
-            {t('paymentTerminals.title')}
-          </h1>
-          <p className="mt-1 text-sm text-slate-500">{t('paymentTerminals.description')}</p>
-        </div>
+    <div className="max-w-3xl space-y-6">
+      {/* Register button — page header chrome lives with the caller now
+          (this panel is embedded in the branch hub's Terminals tab), but the
+          register action must stay reachable. */}
+      <div className="flex justify-end">
         <Button onClick={() => setIsFormOpen(true)} className="flex-shrink-0">
           <Plus className="mr-1.5 h-4 w-4" aria-hidden="true" />
           {t('paymentTerminals.register')}
         </Button>
       </div>
 
-      <div className="max-w-3xl space-y-6">
-        {/* Honest fail-closed note */}
-        <div className="flex items-start gap-3 rounded-xl border border-amber-200 bg-amber-50 p-4 text-sm text-amber-800">
-          <ShieldCheck className="mt-0.5 h-5 w-5 flex-shrink-0 text-amber-500" aria-hidden="true" />
-          <span>{t('paymentTerminals.gateNote')}</span>
-        </div>
-
-        {/* Terminals */}
-        <SettingsSection
-          title={t('paymentTerminals.sectionTitle')}
-          description={t('paymentTerminals.sectionDescription')}
-          icon={<CreditCard className="h-4 w-4" />}
-        >
-          {isLoading ? (
-            <div className="flex justify-center py-6">
-              <Spinner />
-            </div>
-          ) : !terminals || terminals.length === 0 ? (
-            <div className="flex flex-col items-center gap-3 py-8 text-center">
-              <div className="rounded-full bg-slate-100 p-3 text-slate-400">
-                <CreditCard className="h-6 w-6" aria-hidden="true" />
-              </div>
-              <p className="text-sm text-slate-500">{t('paymentTerminals.empty')}</p>
-              <Button size="sm" variant="secondary" onClick={() => setIsFormOpen(true)}>
-                <Plus className="mr-1.5 h-4 w-4" aria-hidden="true" />
-                {t('paymentTerminals.register')}
-              </Button>
-            </div>
-          ) : (
-            <ul className="divide-y divide-slate-100">
-              {terminals.map((rec) => (
-                <li
-                  key={rec.id}
-                  className="flex flex-wrap items-center justify-between gap-3 py-3 first:pt-0 last:pb-0"
-                >
-                  <div className="min-w-0">
-                    <div className="flex flex-wrap items-center gap-2">
-                      <span className="font-medium text-slate-900">{providerLabel(rec.providerId)}</span>
-                      <Badge variant={STATE_VARIANT[rec.activationState]} size="sm">
-                        {t(`paymentTerminals.state.${rec.activationState}`)}
-                      </Badge>
-                      {rec.fiscalCoupled && (
-                        <Badge variant="primary" size="sm">
-                          {t('paymentTerminals.fiscalCoupled')}
-                        </Badge>
-                      )}
-                      {!rec.providerRegistered && (
-                        <Badge variant="danger" size="sm">
-                          {t('paymentTerminals.providerMissing')}
-                        </Badge>
-                      )}
-                    </div>
-                    <div className="mt-1 truncate text-sm text-slate-500">
-                      {t('paymentTerminals.serial')}: {rec.serial}
-                      {rec.model ? ` · ${rec.model}` : ''}
-                      {rec.deviceId ? ` · ${t('paymentTerminals.device')}` : ''}
-                    </div>
-                  </div>
-                  <div className="flex flex-shrink-0 items-center gap-2">
-                    {renderActions(rec)}
-                    <button
-                      onClick={() => removeTerminal(rec)}
-                      aria-label={t('paymentTerminals.remove')}
-                      className="rounded-lg p-2 text-slate-400 transition-colors hover:bg-red-50 hover:text-red-600"
-                    >
-                      <Trash2 className="h-4 w-4" aria-hidden="true" />
-                    </button>
-                  </div>
-                </li>
-              ))}
-            </ul>
-          )}
-        </SettingsSection>
-
-        {/* Reconciliation — only when there's something to act on. */}
-        {reconciliation && reconciliation.length > 0 && (
-          <section className="overflow-hidden rounded-xl border border-red-200 bg-white">
-            <div className="flex items-start gap-3 border-b border-red-100 bg-red-50/60 px-5 py-4">
-              <div className="flex-shrink-0 rounded-lg bg-red-100 p-2 text-red-600">
-                <AlertTriangle className="h-4 w-4" aria-hidden="true" />
-              </div>
-              <div>
-                <h3 className="text-base font-semibold text-slate-900">
-                  {t('paymentTerminals.reconciliation.title')}
-                </h3>
-                <p className="mt-0.5 text-sm text-slate-500">
-                  {t('paymentTerminals.reconciliation.description')}
-                </p>
-              </div>
-            </div>
-            <ul className="divide-y divide-slate-100 px-5 py-2">
-              {reconciliation.map((c) => (
-                <li
-                  key={c.chargeId}
-                  className="flex flex-wrap items-center justify-between gap-2 py-3"
-                >
-                  <span className="font-medium text-slate-800">
-                    #{c.orderId.slice(0, 8)} · {formatPrice(c.amount)}
-                  </span>
-                  <span className="flex items-center gap-2">
-                    {c.approvalCode && (
-                      <span className="text-xs text-slate-400">{c.approvalCode}</span>
-                    )}
-                    <Badge variant={c.status === 'NEEDS_REVIEW' ? 'danger' : 'warning'} size="sm">
-                      {t(`paymentTerminals.reconciliation.status.${c.status}`, c.status)}
-                    </Badge>
-                  </span>
-                </li>
-              ))}
-            </ul>
-          </section>
-        )}
+      {/* Honest fail-closed note */}
+      <div className="flex items-start gap-3 rounded-xl border border-amber-200 bg-amber-50 p-4 text-sm text-amber-800">
+        <ShieldCheck className="mt-0.5 h-5 w-5 flex-shrink-0 text-amber-500" aria-hidden="true" />
+        <span>{t('paymentTerminals.gateNote')}</span>
       </div>
+
+      {/* Terminals */}
+      <SettingsSection
+        title={t('paymentTerminals.sectionTitle')}
+        description={t('paymentTerminals.sectionDescription')}
+        icon={<CreditCard className="h-4 w-4" />}
+      >
+        {isLoading ? (
+          <div className="flex justify-center py-6">
+            <Spinner />
+          </div>
+        ) : !terminals || terminals.length === 0 ? (
+          <div className="flex flex-col items-center gap-3 py-8 text-center">
+            <div className="rounded-full bg-slate-100 p-3 text-slate-400">
+              <CreditCard className="h-6 w-6" aria-hidden="true" />
+            </div>
+            <p className="text-sm text-slate-500">{t('paymentTerminals.empty')}</p>
+            <Button size="sm" variant="secondary" onClick={() => setIsFormOpen(true)}>
+              <Plus className="mr-1.5 h-4 w-4" aria-hidden="true" />
+              {t('paymentTerminals.register')}
+            </Button>
+          </div>
+        ) : (
+          <ul className="divide-y divide-slate-100">
+            {terminals.map((rec) => (
+              <li
+                key={rec.id}
+                className="flex flex-wrap items-center justify-between gap-3 py-3 first:pt-0 last:pb-0"
+              >
+                <div className="min-w-0">
+                  <div className="flex flex-wrap items-center gap-2">
+                    <span className="font-medium text-slate-900">{providerLabel(rec.providerId)}</span>
+                    <Badge variant={STATE_VARIANT[rec.activationState]} size="sm">
+                      {t(`paymentTerminals.state.${rec.activationState}`)}
+                    </Badge>
+                    {rec.fiscalCoupled && (
+                      <Badge variant="primary" size="sm">
+                        {t('paymentTerminals.fiscalCoupled')}
+                      </Badge>
+                    )}
+                    {!rec.providerRegistered && (
+                      <Badge variant="danger" size="sm">
+                        {t('paymentTerminals.providerMissing')}
+                      </Badge>
+                    )}
+                  </div>
+                  <div className="mt-1 truncate text-sm text-slate-500">
+                    {t('paymentTerminals.serial')}: {rec.serial}
+                    {rec.model ? ` · ${rec.model}` : ''}
+                    {rec.deviceId ? ` · ${t('paymentTerminals.device')}` : ''}
+                  </div>
+                </div>
+                <div className="flex flex-shrink-0 items-center gap-2">
+                  {renderActions(rec)}
+                  <button
+                    onClick={() => removeTerminal(rec)}
+                    aria-label={t('paymentTerminals.remove')}
+                    className="rounded-lg p-2 text-slate-400 transition-colors hover:bg-red-50 hover:text-red-600"
+                  >
+                    <Trash2 className="h-4 w-4" aria-hidden="true" />
+                  </button>
+                </div>
+              </li>
+            ))}
+          </ul>
+        )}
+      </SettingsSection>
+
+      {/* Reconciliation — only when there's something to act on. */}
+      {reconciliation && reconciliation.length > 0 && (
+        <section className="overflow-hidden rounded-xl border border-red-200 bg-white">
+          <div className="flex items-start gap-3 border-b border-red-100 bg-red-50/60 px-5 py-4">
+            <div className="flex-shrink-0 rounded-lg bg-red-100 p-2 text-red-600">
+              <AlertTriangle className="h-4 w-4" aria-hidden="true" />
+            </div>
+            <div>
+              <h3 className="text-base font-semibold text-slate-900">
+                {t('paymentTerminals.reconciliation.title')}
+              </h3>
+              <p className="mt-0.5 text-sm text-slate-500">
+                {t('paymentTerminals.reconciliation.description')}
+              </p>
+            </div>
+          </div>
+          <ul className="divide-y divide-slate-100 px-5 py-2">
+            {reconciliation.map((c) => (
+              <li
+                key={c.chargeId}
+                className="flex flex-wrap items-center justify-between gap-2 py-3"
+              >
+                <span className="font-medium text-slate-800">
+                  #{c.orderId.slice(0, 8)} · {formatPrice(c.amount)}
+                </span>
+                <span className="flex items-center gap-2">
+                  {c.approvalCode && (
+                    <span className="text-xs text-slate-400">{c.approvalCode}</span>
+                  )}
+                  <Badge variant={c.status === 'NEEDS_REVIEW' ? 'danger' : 'warning'} size="sm">
+                    {t(`paymentTerminals.reconciliation.status.${c.status}`, c.status)}
+                  </Badge>
+                </span>
+              </li>
+            ))}
+          </ul>
+        </section>
+      )}
 
       {/* Register modal */}
       <Modal
@@ -307,5 +301,3 @@ const PaymentTerminalsSettingsPage = () => {
     </div>
   );
 };
-
-export default PaymentTerminalsSettingsPage;
